priliminary moving of unit tests to respective folders
