+++ test/enclave_wrangler/test_actions_api
@@ -1,0 +1,70 @@
+"""Tests
+
+How to run:
+    python -m unittest discover
+
+TODO's
+ - 1. Test framework: Current implementation is ad-hoc for purposes of development.
+ - 2. Change from validate to apply, or do both
+"""
+import os
+import pickle
+import sys
+import unittest
+from datetime import datetime, timedelta
+from pathlib import Path
+from typing import Dict, List, Union
+
+import pandas as pd
+from requests import Response
+from sqlalchemy.exc import IntegrityError
+
+TEST_DIR = os.path.dirname(__file__)
+PROJECT_ROOT = Path(TEST_DIR).parent
+# todo: why is this necessary in this case and almost never otherwise?
+# https://stackoverflow.com/questions/33862963/python-cant-find-my-module
+sys.path.insert(0, str(PROJECT_ROOT))
+
+TEST_INPUT_DIR = os.path.join(TEST_DIR, '../input', 'test_enclave_wrangler')
+TEST_SCHEMA = 'test_n3c'
+yesterday: str = (datetime.now() - timedelta(days=1)).isoformat() + 'Z'  # works: 2023-01-01T00:00:00.000Z
+
+from enclave_wrangler.actions_api import upload_concept_set_version_draft
+
+class TestActionsApi(unittest.TestCase):
+    def test_upload_concept_set_version(self):
+        response: Response = upload_concept_set_version_draft(
+            domain_team='x', provenance='x', current_max_version=2.1, concept_set='x', annotation='x', limitations='x',
+            intention='x', base_version=1, intended_research_project='x', version_id=1, authority='x')
+        # self.assertTrue('result' in response and not response['result'] == 'VALID')
+        self.assertLess(response.status_code, 400)
+
+    # TODO: teardown: "apiName": "archive-concept-set", (for version or container?)
+    #     {
+    #       "apiName": "archive-concept-set",
+    #       "description": "Sets Concept Set 'archived' property to true so that it no longer appears in browser",
+    #       "rid": "ri.actions.main.action-type.cbc3643b-cca4-4772-ae2d-ae7036a6798b",
+    #       "parameters": {
+    #         "concept-set": {
+    #           "description": "",
+    #           "baseType": "OntologyObject"
+    #         }
+    #       }
+    #     },
+
+    # TODO: teardown: "apiName": "delete-omop-concept-set-version",
+    #     {
+    #       "apiName": "delete-omop-concept-set-version",
+    #       "description": "Do not forget to the flag 'Is Most Recent Version' of the previous Concept Set Version to True!",
+    #       "rid": "ri.actions.main.action-type.93b82f88-bd55-4daf-a0f9-f6537bf2bce1",
+    #       "parameters": {
+    #         "omop-concept-set": {
+    #           "description": "",
+    #           "baseType": "OntologyObject"
+    #         },
+    #         "expression-items": {
+    #           "description": "",
+    #           "baseType": "Array<OntologyObject>"
+    #         }
+    #       }
+    #     },
